Move arena source file to fem dir.
- split definition of classes to two header files
